Exterminated some pesty bugs
to-do: go for a walk lol
