+++ frontend/src/pages/exams/ExamsView.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
-import HeaderNav from "../../components/NestedList";
+import HeaderNav from "../../components/HeaderNav";
 import ExamCard from "../../components/ExamCard";
 import "./exams.scss";
