+++ frontend/src/pages/schedule/Calendar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Slider, Grid, Row, Column, Loading } from "carbon-components-react";
-import { addDays, format, formatISO } from "date-fns";
+import { addDays, format } from "date-fns";
 import Day from "./Day";
 import api from "../../apiHandle";
 
@@ -46,10 +46,6 @@ class Calendar extends React.Component {
     return this.state.days > 0;
   }
 
-  hello() {
-    console.log("I was triggered during render");
-  }
-
   componentDidMount() {
     console.log(this.props);
     // this.interval = setInterval(() =>
@@ -60,6 +56,10 @@ class Calendar extends React.Component {
     });
     // );
   }
+
+  handleCallback = (data) => {
+    this.setState(data);
+  };
 
   // componentWillUnmount() {
   //   clearInterval(this.interval);
@@ -97,10 +97,7 @@ class Calendar extends React.Component {
         });
         this.setState({ expand: "no-expand" });
       }
-
-      // setexpand = (wrapper.expand === "no-expand") ? 'is-expanded' : 'no-expand'
     }
-    // console.log(wrapper.expand);
   }
 
   //goes through Day components and finds out which index day is expanded
@@ -118,16 +115,16 @@ class Calendar extends React.Component {
   }
 
   changeDayVal(value) {
+    console.log(this.state.expand)
     if (this.state.expand === "is-expanded") {
       let day = this.findExpandedDay();
       console.log(day);
       this.expandDay(day, null);
-
       this.state.expand === "is-expanded" && this.setState({ days: value });
     }
     console.log(this.state.expand);
-
-    this.state.expand === "no-expand" && this.setState({ days: value });
+    console.log(value);
+    this.setState({expand: "no-expand", days: value});
   }
 
   render() {
@@ -139,32 +136,26 @@ class Calendar extends React.Component {
             <Day
               key={i}
               i={i}
-              // className="today"
               today={true}
               fullDate={this.state.today}
-              month={format(this.state.today, "LLL")}
               date={this.state.today.getDate()}
               day={format(this.state.today, "EEE")}
               days={this.state.days}
-              ref={(ref) => (this.daysrefs[i] = ref)}
               expand="no-expand"
-              // expandWidth={document.getElementById('calendar').clientWidth}
+              parentCallback={this.handleCallback}
               user={this.state.user}
             />
           ) : (
             <Day
               key={i}
               i={i}
-              // className={this.state.expand}
               today={false}
               fullDate={addDays(this.state.today, i)}
-              month={format(addDays(this.state.today, i), "LLL")}
               date={addDays(this.state.today, i).getDate()}
               day={format(addDays(this.state.today, i), "EEE")}
               days={this.state.days}
-              ref={(ref) => (this.daysrefs[i] = ref)}
               expand="no-expand"
-              // expandWidth={document.getElementById('calendar').clientWidth}
+              parentCallback={this.handleCallback}
               user={this.state.user}
             />
           );
@@ -194,16 +185,14 @@ class Calendar extends React.Component {
             <h4> {this.state.days} days </h4>
             <div className="slider-row">
               <div>
-                <div>
-                  <Slider
-                    ariaLabelInput="day slider"
-                    className="day-slider"
-                    max={5}
-                    min={1}
-                    value={3}
-                    onChange={({ value }) => this.changeDayVal(value)}
-                  />
-                </div>
+                <Slider
+                  ariaLabelInput="day slider"
+                  className="day-slider"
+                  max={5}
+                  min={1}
+                  value={3}
+                  onChange={({ value }) => this.changeDayVal(value)}
+                />
               </div>
             </div>
           </Column>
@@ -211,16 +200,13 @@ class Calendar extends React.Component {
 
         {this.state.dataLoaded ? (
           <Row className="date-head">
-            {day_list.map((day, index) => {
-              let handleclick = this.expandDay.bind(this, index);
+            {day_list.map((day) => {
 
               return (
                 <Column
                   key={day.props.date}
-                  // className={'a-day bx--col-lg-' + Math.floor(16 / this.state.days)}
                   className="a-day"
                   lg={Math.floor(16 / this.state.days)}
-                  // onClick={handleclick}
                 >
                   {day}
                 </Column>
