+++ frontend/src/pages/schedule/Day.js
@@ -1,19 +1,29 @@
 import React from "react";
-import PropTypes from "prop-types";
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbSkeleton,
-} from "carbon-components-react";
 import { ChevronLeft32 } from "@carbon/icons-react";
-import Breadcrumbs from "@material-ui/core/Breadcrumbs";
-import LinkUI from "@material-ui/core/Link";
 import { Link } from "react-router-dom";
 import api from "../../apiHandle";
 import { parseISO, getHours, getMinutes, isSameDay, format } from "date-fns";
 import LectureCard from "../../components/LectureCard";
 import AssignmentCard from "../../components/AssignmentCard";
 import ExamCard from "../../components/AssignmentCard";
+import SyllabusView from './components/SyllabusView';
+
+function BackButton(props) {
+
+  return (
+    <button
+      className="icon-text"
+      onClick={() => props.parentCallback({
+        expandedCourse: null,
+        syllabusView: false,
+        expand: "no-expand",
+      })}
+    >
+      <ChevronLeft32 aria-label="backSchedule" className="back-button" />
+      <p>Schedule</p>
+    </button>
+  );
+}
 
 function Badge(props) {
   if (props.type === "assignments") {
@@ -112,167 +122,6 @@ function DayHeader(props) {
     );
   }
 }
-
-function SyllabusView(props) {
-  var right_side_css = {
-    display:
-      props.expand === "is-expanded" && props.syllabusView ? "block" : "none",
-  };
-
-  const { course, type } = props;
-  console.log(props);
-  console.log(course.courseId);
-
-  switch (type) {
-    case "assignments":
-      return (
-        <div className="right-side" style={right_side_css}>
-          <div style={{ display: "flex", flexDirection: "row" }}>
-            <div className="syllabus-column">
-              <h6 style={{ color: "gray" }}>
-                {`${course.courseId} - ${course.courseName} ${course.section}`}
-              </h6>
-              <br></br>
-
-              <h4>Assignments:</h4>
-              <br></br>
-
-              {course.assignments.length == 0 ? (
-                <p>No assignments for this course.</p>
-              ) : (
-                course.assignments.map((a) => {
-                  return <AssignmentCard {...a} />;
-                })
-              )}
-            </div>
-          </div>
-        </div>
-      );
-
-    case "exams":
-      return (
-        <div className="right-side" style={right_side_css}>
-          <div style={{ display: "flex", flexDirection: "row" }}>
-            <div className="syllabus-column">
-              <h6 style={{ color: "gray" }}>
-                {`${course.courseId} - ${course.courseName} ${course.section}`}
-              </h6>
-
-              <br></br>
-              <h4>
-                All exams for{" "}
-                {<span style={{ color: "orange" }}>{course.courseId}</span>}:
-              </h4>
-
-              {course.exams.length == 0 ? (
-                <p>No exams for this course.</p>
-              ) : (
-                course.exams.map((e) => {
-                  return <ExamCard {...e} />;
-                })
-              )}
-            </div>
-          </div>
-        </div>
-      );
-
-    case "course-overview":
-      return (
-        <div className="right-side" style={right_side_css}>
-          <div style={{ display: "flex", flexDirection: "row" }}>
-            <div className="syllabus-column">
-              <h5 style={{ color: "gray" }}>
-                {`${course.courseId} - ${course.courseName} ${course.section}`}
-              </h5>
-              <br></br>
-
-              <h4>Office Hours:</h4>
-              <br></br>
-
-              <h5>{`Professor ${course.professor[0].firstName} ${course.professor[0].lastName}: Wednesday 10am-11am`}</h5>
-              <div className="zoomlink">
-                <a
-                  className="blue"
-                  target="_blank"
-                  rel="noreferrer"
-                  // href={props.zoomLink}
-                >
-                  <h5>Zoom Link</h5>
-                </a>
-              </div>
-              <br></br>
-
-              <h5>{`TA Fake TA-Man: Monday 12pm-1pm`}</h5>
-              <div className="zoomlink">
-                <a
-                  className="blue"
-                  target="_blank"
-                  rel="noreferrer"
-                  // href={props.zoomLink}
-                >
-                  <h5>Zoom Link</h5>
-                </a>
-              </div>
-              <br></br>
-
-              <h4>
-                Due today:{" "}
-                <button
-                  className="see-all"
-                  onClick={() =>
-                    props.parentCallback({ syllabusViewType: "assignments" })
-                  }
-                >
-                  <h4>See all</h4>
-                </button>
-              </h4>
-              <br></br>
-              {course.assignments.length == 0 ? (
-                <p>No assignments due today.</p>
-              ) : (
-                course.assignments.map((a) => {
-                  return (
-                    <button key={a.id} className={"assignment-small"}>
-                      <div>
-                        <h5>{a.name}</h5>
-                      </div>
-                    </button>
-                  );
-                })
-              )}
-
-              <h4>
-                Exams today:{" "}
-                <button
-                  className="see-all"
-                  onClick={() =>
-                    props.parentCallback({ syllabusViewType: "exams" })
-                  }
-                >
-                  <h4>See all</h4>
-                </button>
-              </h4>
-              <br></br>
-              {course.exams.length == 0 ? (
-                <p>No exams for this course.</p>
-              ) : (
-                course.exams.map((e) => {
-                  return (
-                    <button key={e.id} className={"exam-small"}>
-                      <div>
-                        <h5>{e.name}</h5>
-                      </div>
-                    </button>
-                  );
-                })
-              )}
-            </div>
-          </div>
-        </div>
-      );
-  }
-}
-
 class Day extends React.Component {
   constructor(props) {
     super(props);
@@ -280,23 +129,33 @@ class Day extends React.Component {
       lectures: [],
       labs: [],
       assignments: [],
-      all: [],
       exams: [],
       expand: "no-expand", //classname to add to day component to indicate css transition
       prevExpand: "no-expand", //set previous state of expansion to prevent repeated events into expanded view
       expandedCourse: null,
       syllabusView: false,
       syllabusViewType: "course-overview",
-      rightSide: {},
       colWidth: 0,
     };
     this.test = false;
-    this.courserefs = []; //array of refs to data in a lecture card
   }
 
   componentDidUpdate() {
-    console.log("f");
-    return true;
+    if (this.state.expand === "is-expanded") {
+      let width = this.getEWidth();
+      width !== this.state.colWidth &&
+        this.setState({
+          colWidth: width,
+        });
+    }
+  }
+
+  backButtonCallback() {
+    this.setState({
+      expandedCourse: null,
+      syllabusView: false,
+      expand: "no-expand",
+    })
   }
 
   componentDidMount() {
@@ -334,6 +193,8 @@ class Day extends React.Component {
       });
     });
 
+    console.log(lectures.concat(labs).sort((l1, l2) => l1.startTime < l2.startTime));
+
     this.setState({
       lectures: lectures,
       labs: labs,
@@ -355,12 +216,6 @@ class Day extends React.Component {
     );
   };
 
-  setLectureRef = (el) => {
-    this.expandRef = el;
-    this.daysrefs.push(this.showMoreRef);
-    console.log(el);
-  };
-
   handleCallback = (courseId) => {
     console.log(courseId);
     // this.setState(childData)
@@ -370,21 +225,18 @@ class Day extends React.Component {
     // If day is expanded AND a course's syllabus view is open
     //    If clicked course is currently open course, close syllabusview
     if (this.state.expand === "is-expanded" && this.state.syllabusView) {
-      console.log("expanded and syllabus");
-      toExpand.courseId !== this.state.expandedCourse.courseId
-        ? this.setState({
-            expandedCourse: toExpand,
-            syllabusViewType: "course-overview",
-          })
-        : this.setState({ expandedCourse: null, syllabusView: false });
-
+      if (toExpand.courseId !== this.state.expandedCourse.courseId) {
+        this.setState({
+          expandedCourse: toExpand,
+          syllabusViewType: "course-overview",
+        })
+      } else { 
+        this.setState({ expandedCourse: null, syllabusView: false, expand: "no-expand"});
+      }
+        
       // If day is expanded AND no syllabus view is open
       //    Open course's syllabus view
-    } else if (
-      this.state.expand === "is-expanded" &&
-      !this.state.syllabusView
-    ) {
-      console.log("expanded and !syllabus");
+    } else if (this.state.expand === "is-expanded" && !this.state.syllabusView) {
       this.setState({
         syllabusView: true,
         expandedCourse: toExpand,
@@ -394,7 +246,6 @@ class Day extends React.Component {
       // If day is not expanded
       //    Expand day and open clicked course's syllabus view
     } else if (this.state.expand === "no-expand") {
-      console.log("this happened4");
       this.setState({
         expand: "is-expanded",
         syllabusViewType: "course-overview",
@@ -406,43 +257,18 @@ class Day extends React.Component {
 
   childCallback = (data) => {
     this.setState(data);
+    this.props.parentCallback(data);
   };
-
-  //On-Click function that shows right hand side expanded view for class on
-  //course card clikck
-  showMore(index, e, course) {
-    console.log(course);
-    if (this.state.expand === "is-expanded") {
-      // console.log(index)
-      // let lecture = this.courserefs[index]
-      // console.log(this.state.lectures[index])
-      // this.populateRightSide(lecture)
-      // // console.log("^ from show more")
-
-      e.stopPropagation();
-      this.setState((prevState) => {
-        return {
-          expand: "is-expanded",
-          prevExpand: prevState.expand,
-        };
-      });
-      this.setState({ syllabusView: !this.state.syllabusView });
-    }
-  }
 
   getEWidth() {
     let elem;
     if (this.props.days > 1) {
       elem = document.querySelector(".no-expand");
     }
-    let width = this.props.days > 1 ? elem.clientWidth : 0;
-    return width;
+    return this.props.days > 1 ? elem.clientWidth : 0;
   }
 
   render() {
-    // console.log("expad: " + this.state.expand)
-    // console.log("prev expand: " + this.state.prevExpand)
-    // console.log("colWid: " + this.state.colWidth)
 
     var { lectures, labs } = this.state;
 
@@ -450,10 +276,10 @@ class Day extends React.Component {
       this.state.expand === "is-expanded"
         ? {
             transform: `translateX(calc(-${
-              this.state.colWidth * this.props.i
+              this.state.colWidth * (this.props.i)
             }px - calc(1rem *  ${this.props.i})))`,
           }
-        : { transform: `translateX(0px)` };
+        : { transform: `translateX(0px)`};
 
     return (
       <section
@@ -466,22 +292,8 @@ class Day extends React.Component {
         style={section_css}
       >
         {this.state.expand === "is-expanded" ? (
-          <button
-            className="icon-text"
-            onClick={() =>
-              this.setState({
-                expandedCourse: null,
-                syllabusView: false,
-                expand: "no-expand",
-              })
-            }
-          >
-            <ChevronLeft32 aria-label="backSchedule" className="back-button" />
-            <p>Schedule</p>
-          </button>
-        ) : (
-          false
-        )}
+          <BackButton parentCallback={this.childCallback}/>
+        ) : (false)}
 
         <DayHeader
           expand={this.state.expand}
@@ -499,7 +311,7 @@ class Day extends React.Component {
               type="exams"
               exams={this.state.exams}
               user={this.props.user}
-            ></Badge>
+            />
             {/* // Assignments badge */}
             <Badge
               type="assignments"
@@ -514,8 +326,8 @@ class Day extends React.Component {
             {lectures.length + labs.length !== 0 ? (
               lectures
                 .concat(labs)
-                .sort((l1, l2) => l1.startTime > l2.startTime)
-                .map((course, i) => {
+                .sort((l1, l2) => l1.id > l2.id)
+                .map((course) => {
                   return (
                     <LectureCard
                       key={course.courseId}
@@ -529,7 +341,6 @@ class Day extends React.Component {
                       showFull={this.props.days <= 4} //show full zoom link when schedule is on 3-day view and below
                       parentCallback={this.handleCallback}
                       syllabusView={this.state.syllabusView} //bool to toggle right-hand side details
-                      // ref={(ref) => (this.courserefs[i] = ref)}
                     />
                   );
                 })
@@ -550,9 +361,7 @@ class Day extends React.Component {
               type={this.state.syllabusViewType}
               parentCallback={this.childCallback}
             />
-          ) : (
-            false
-          )}
+          ) : (false)}
         </div>
       </section>
     );
